Made small changes to IndependentSolver::DenseSet and IndependentSolver::IndpendentElment.
Specifically, the ability to record which ref<Expr> are in a
particular factor are now recorded.  This is important for
future changes to IndependentSolver::getInitialValues() so
that an incoming constraint can be broken down into it's
smallest possible parts.  Each of these individual parts may
then be solved for and then the solutions to each piece
combined to create a final answer.

Finally, several fields which had previously been private are now
public to facilitate the smaller solutions being combined into a
larger solution.