Added the ability to solve for all factors in a particular query.
This functionality is necessary in order to more effectively handle
calls to IndependentSolver::getInitialValues.  An incoming query will
be broken down into its smaller parts, and each piece will be solved
for. At the end, the pieces will be recombined into a larger solution.
The IndependentElementSet::getAllFactors() method takes a query and
breaks it down into all of it's non-interacting factors.  The
IndependentElementSet::calculateArrays() method calculates which
arrays are involved in a particular factor.

Conflicts:
	lib/Solver/IndependentSolver.cpp