Commit of improved IndependentSolver::getIniitalValues(). The previous implementation simply passes the entire constraint forward without any factoring of the constraint at all.  This is a problem since it is highly likely that there are cached solutions to pieces of the constraint.  The new implementation breaks the entire constraint down into its requisite factors and passes each piece forward, one by one, down the solver chain.  After an answer is returned, it is integrated into a larger solution.  Since, by definition, no factor can affect another, we can safely create a solution to the larger constraint from the answers of its smaller pieces.
The reconstruction of the solution is done by analyzing which parts of
an array a factor touches.  If the factor is the only one to reference
a particular array, then all of the values calculated in the solution
for that array are included in the final answer.  If the factor
references a particular element of the array (for example, arr[1]),
then only the value in index 1 of array arr will be included in the
solution.