Added a max-forks-terminate option to the commandline
The max-forks-terminate option kills exploration after reaching a
certain number of forks.  Klee then begins to dump states and wrap up
execution.  This metric is similar to the current max-forks commandline
option.  The main difference is that the dangling states are not pushed
through the rest of the pogram.

The anticipated use case for this option is to terminate after a
certain number of forks and then create tests for each of the dangling
states.  Each of these tests will then be executed normally, rather
than inside klee, to make sure there are no assertion violations or
errors.  This is likely MUCH faster than finishing each path within
Klee (like max-forks does).