Improved MaxForks by using cached solutions instead of SAT calls
This commit optimizes programs whose execution is halted by the
MaxForks option.  Once a particular number of forks is hit, the
execution stops forking.  It does not, however, stop exploring the
program. Instead each "dangling" state is pushed to completion to
make sure no assertion violations or errors lie along the path.

Previously, the implementation had used SAT calls to test the validity
of a branch.  If both sides were possible, it would randomly pick one
side to push the execution down.

The optimization works by calculating a solution to the dangling state
and following that solution through to completion.  This is comparable
to generating a test case for the dangling state in klee, exiting klee,
and running the test case as you would any other input.  This means
that any problems that lie further down the path will be uncovered. By
avoiding the sat calls and instead relying on simply plugging a
solution into a state and seeing if it is T/F, the program completes
much more quickly.