This commit tests whether the point that satisfied the parent of a new state will also satisfy the new state itself.
The opportunity for this optimization occurs every time a state is
forked.  This is because it is guarenteed the model (solution/
assignment) that satisfies the parent state will also satisfy one of
its two children.  Therefore, this change short-circuits the UBTree
present in CexCachingSolver by seeing whether the solution to the
parent of some newly generated state satisfies the new state as well.
It is guarenteed to work at least 50% of the time so it's a chance
worth taking.  Only if this check fails, i.e. the parent's solution
does not satisfy the new state, will the UBTree be searched.