Added logic to simplifier that creates equality out of > and <
This commit adds the ability to concretize a variable's value by
using inequalities to find the range of values a variable can be.
Should the range be reduced to a single value, then the variable
can be concretized. For example if the state contains the constraints
(x>6) and (x<=7) then we know that (x==7).  This equality can then be
used to simplify the incoming query.expr to a smaller and more
manageable form.  Sometimes the query.expr can be simplified to a
concrete value, eliminating the need for a Solver call at all.